requirements.txt
numpy>=1.21.0
trimesh>=3.9.0
pyrender>=0.1.45
Pillow>=8.3.0
imageio>=2.19.0
opencv-python>=4.5.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
pyopengl>=3.1.6
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.0
pyyaml>=5.4.0
tqdm>=4.62.0
pandas>=1.3.0
scipy>=1.7.0
